refactor(html): keep the frontend css and js as files, embedded at build time
`frontend.cpp` was 92 KB, 78 KB of it css and javascript inside raw string
literals. Nothing that reads either language could see it: no highlighting, no
formatter, no linter, and the three `test/browser` harnesses each parsed the C++
by string index to recover a script and wrote a gitignored copy beside their
pages. msvc caps a string literal at 16380 bytes, so `pdf_annotation_js` was
already split over two literals reassembled through an `Asset::content_tail`,
guarded by a `consteval fits_a_literal`, and `spreadsheet.js` has since been
split the same way.

The 21 assets now live as files under `src/odr/internal/html/frontend/`, and
`cmake/frontend_assets.cmake` embeds them into a generated header of
`inline constexpr std::string_view`s. The bytes back them as a `char` array of
`'\xNN'` character literals: an array has no length cap, and the literal form
survives both a signed and an unsigned `char` — a `0xNN` integer narrows on one
and a negative one on the other.

This is not the `data/` directory #648 removed. Nothing is shipped beside the
library, no consumer gains a path to point at, and android, apple, wasm, python
and the jni all build the same target as before. `frontend.cpp` keeps the asset
table, which now reads as the documented index of what each view writes, and
loses `content_tail` and `fits_a_literal`.

The emitted html is byte-identical: `HtmlOutputTests` reproduces every reference
output, `resources/*.css` and `*.js` included. The newline that used to open
each literal — and that a shipped resource file still carries — is written once
now, in `written()`, so the two forms stay the same bytes. `pdf-annotation.js`
keeps the blank line at the seam of the two old literals for the same reason.

`test/browser/*/serve` no longer extract anything: a shared `serve.py` falls
back to the asset directory, so a page links the file the library embeds.
