+++ src/odr/internal/html/frontend/document.js
@@ -1,0 +1,52 @@
+(function () {
+  "use strict";
+
+  var odr = (window.odr = window.odr || {});
+
+  odr.onError = function (code, message) {
+    console.error("error " + code + " message " + message);
+  };
+
+  var errorIllegalEditNewLine = {
+    code: 1,
+    message: "new line not supported by this document",
+  };
+
+  var modified = {};
+
+  odr.generateDiff = function () {
+    var ops = [];
+    for (var path in modified) {
+      if (Object.prototype.hasOwnProperty.call(modified, path)) {
+        ops.push({ op: "setText", path: path, text: modified[path].innerText });
+      }
+    }
+    return JSON.stringify({ version: 1, ops: ops });
+  };
+
+  new MutationObserver(function (mutations) {
+    for (var i = 0; i < mutations.length; ++i) {
+      if (mutations[i].type !== "characterData") {
+        continue;
+      }
+      // The nearest owner, not the direct parent: a search `<mark>` may sit
+      // between the edited text and the element carrying the path.
+      var parent = mutations[i].target.parentElement;
+      var owner = parent && parent.closest("[data-odr-path]");
+      if (owner) {
+        modified[owner.getAttribute("data-odr-path")] = owner;
+      }
+    }
+  }).observe(document.body, {
+    childList: true,
+    subtree: true,
+    characterData: true,
+  });
+
+  document.addEventListener("keydown", function (event) {
+    if (event.key === "Enter") {
+      event.preventDefault();
+      odr.onError(errorIllegalEditNewLine.code, errorIllegalEditNewLine.message);
+    }
+  });
+})();
