+++ src/odr/internal/html/frontend/media.css
@@ -1,0 +1,4 @@
+body{margin:0;background:#000}
+.odr-media{display:flex;align-items:center;justify-content:center;min-height:100vh}
+.odr-media video{max-width:100%;max-height:100vh}
+.odr-media audio{width:100%;max-width:40rem;margin:0 1rem}
