+++ src/odr/internal/html/frontend/search-dark.css
@@ -1,0 +1,1 @@
+mark{color:#0d1117!important}
