+++ src/odr/internal/html/frontend/search.css
@@ -1,0 +1,2 @@
+mark{background:#ff0}
+mark.current{background:orange}
