--- test/browser/annotation/.gitignore
@@ -1,2 +1,0 @@
-pdf-annotation.js
-pdf-annotation.css
