--- test/browser/sheet/.gitignore
@@ -1,4 +1,0 @@
-document.css
-spreadsheet.css
-spreadsheet.js
-sheet-editing.js
