+++ test/browser/sheet/serve
@@ -1,50 +1,12 @@
 #!/usr/bin/env python3
-"""Extracts the emitted sheet stylesheet and script and serves the checks."""
+"""Serves the sheet checks."""
 
-import functools
-import http.server
 import pathlib
-import socketserver
-
-PORT = 8732
+import sys
 
 HERE = pathlib.Path(__file__).resolve().parent
-SOURCE = HERE.parents[2] / "src" / "odr" / "internal" / "html" / "frontend.cpp"
-# A script over the 16380 bytes msvc holds in one literal is written as two.
-PARTS = {
-    "document.css": ('constexpr std::string_view document_css = R"css(',),
-    "spreadsheet.css": ('constexpr std::string_view spreadsheet_css = R"css(',),
-    "spreadsheet.js": (
-        'constexpr std::string_view spreadsheet_js = R"js(',
-        'constexpr std::string_view spreadsheet_js_tail = R"js(',
-    ),
-    "sheet-editing.js": ('constexpr std::string_view sheet_editing_js = R"js(',),
-}
+sys.path.insert(0, str(HERE.parent))
 
-PAGES = ("tests.html", "positions.html", "sorting.html", "editing.html")
+from serve import serve  # noqa: E402
 
-
-def extract(source: str, begin: str) -> str:
-    at = source.index(begin)
-    end = ')css";' if begin.endswith('R"css(') else ')js";'
-    return source[at + len(begin) : source.index(end, at)]
-
-
-def main() -> None:
-    source = SOURCE.read_text()
-    for name, parts in PARTS.items():
-        (HERE / name).write_text("".join(extract(source, p) for p in parts))
-    print(f"{SOURCE.name} -> {', '.join(PARTS)}")
-
-    handler = functools.partial(
-        http.server.SimpleHTTPRequestHandler, directory=str(HERE)
-    )
-    socketserver.TCPServer.allow_reuse_address = True
-    with socketserver.TCPServer(("127.0.0.1", PORT), handler) as server:
-        for page in PAGES:
-            print(f"http://localhost:{PORT}/{page}")
-        server.serve_forever()
-
-
-if __name__ == "__main__":
-    main()
+serve(8732, HERE, ("tests.html", "positions.html", "sorting.html", "editing.html"))
