--- test/browser/viewport/.gitignore
@@ -1,1 +1,0 @@
-viewport.js
